moved landing page axios to App
